requirements.txt
Django>=4.2
psycopg2-binary>=2.9
WeasyPrint>=60.0
pytest-django>=4.5
model-bakery>=1.12
python-dateutil>=2.8.2
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv


cloudinary
django-cloudinary-storage
